+++ src/app.ts
@@ -30,23 +30,14 @@ export class App {
   }
 
   run(name: string, ...args): number {
-    const subscribers = this._events[name] || [];
-
+    const subscribers = this.getSubscribers(name, this._events);
     console.assert(subscribers && subscribers.length > 0, 'No subscriber for event: ' + name);
-
-    // Update the list of subscribers by pulling out those which will run once.
-    // We must do this update prior to running any of the events in case they
-    // cause additional events to be turned off or on.
-    this._events[name] = subscribers.filter((sub) => {
-      return !sub.options.once;
-    });
-
     subscribers.forEach((sub) => {
       const { fn, options } = sub;
       if (options.delay) {
         this.delay(name, fn, args, options);
       } else {
-        fn.apply(this, args);
+        Object.keys(options).length > 0 ? fn.apply(this, [...args, options]) : fn.apply(this, args);
       }
       return !sub.options.once;
     });
@@ -62,8 +53,36 @@ export class App {
     if (options._t) clearTimeout(options._t);
     options._t = setTimeout(() => {
       clearTimeout(options._t);
-      fn.apply(this, args);
+      Object.keys(options).length > 0 ? fn.apply(this, [...args, options]) : fn.apply(this, args);
     }, options.delay);
+  }
+
+  query(name: string, ...args): Promise<any[]> {
+    const subscribers = this.getSubscribers(name, this._events);
+    console.assert(subscribers && subscribers.length > 0, 'No subscriber for event: ' + name);
+    const promises = subscribers.map(sub => {
+      const { fn, options } = sub;
+      return Object.keys(options).length > 0 ? fn.apply(this, [...args, options]) : fn.apply(this, args);
+    });
+    return Promise.all(promises);
+  }
+
+  private getSubscribers(name: string, events) {
+    const subscribers = events[name] || [];
+
+    // Update the list of subscribers by pulling out those which will run once.
+    // We must do this update prior to running any of the events in case they
+    // cause additional events to be turned off or on.
+    events[name] = subscribers.filter((sub) => {
+      return !sub.options.once;
+    });
+    Object.keys(events).filter(evt => evt.endsWith('*') && name.startsWith(evt.replace('*', '')))
+      .sort((a, b) => b.length - a.length)
+      .forEach(evt => subscribers.push(...events[evt].map(sub => ({
+        ...sub,
+        options: { ...sub.options, event: name }
+      }))));
+    return subscribers;
   }
 }